Refactor graph module: separate Edge, GraphBuilder, and ImbalancedNodeSet into distinct files for improved organization and maintainability
